backend-python/requirements.txt
fastapi
uvicorn
openai
python-dotenv
httpx
google-generativeai
